pyproject.toml
[tool.poetry]
name = "extractor-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
langchain = "^0.3.13"
sanic = {extras = ["ext"], version = "^24.6.0"}
jinja2 = "^3.1.5"
pydantic = "^2.10.4"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}
langchain-community = "^0.3.13"
langchain-openai = "^0.2.14"
grandalf = "^0.8"

[tool.poetry.group.dev.dependencies]
yoyo-migrations = "^9.0.0"
psycopg2-binary = "^2.9.10"
python-dotenv = {extras = ["cli"], version = "^1.0.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
